requirements.txt
discord.py>=2.0.0
datasets
sympy
antlr4-python3-runtime==4.11
jishaku
python-dotenv
matplotlib
numpy